Generic refactoring of cache
