+++ packages/relative-values/src/app/interfaces/[id]/layout.tsx
@@ -1,5 +1,4 @@
 "use client";
-import { allInterfaces } from "../../../../models/src";
 import { InterfaceProvider } from "@/components/Interface/InterfaceProvider";
 import { Toolbar } from "@/components/Interface/Toolbar";
 import { StyledTab } from "@/components/ui/StyledTab";
