+++ packages/relative-values/src/values/RVStorage.ts
@@ -1,21 +1,9 @@
-import { result } from "@quri/squiggle-lang";
-import { RelativeValue } from "./RelativeValue";
-import { ModelData, distToRelativeValue } from "./SCache";
-
-export type RelativeValueResult = result<RelativeValue, string>;
+import { RelativeValueResult } from "./RelativeValue";
+import { ModelData, searchRelativeValues } from "./RVCache";
 export class RVStorage {
   constructor(public db: ModelData) {}
 
   compare(id1: string, id2: string): RelativeValueResult {
-    if (!this.db.relativeValues[id1] || !this.db.relativeValues[id1][id2]) {
-      return {
-        ok: false,
-        value: `Combination ${id1} and ${id2} not cached`,
-      };
-    }
-    return {
-      ok: true,
-      value: distToRelativeValue(this.db.relativeValues[id1][id2].value),
-    };
+    return searchRelativeValues(this.db, id1, id2)
   }
 }
